VersionTWo!
Turns out one can bind whole hotkeys using keyboard.add_hotkey
Turns out that the best "lookup" time is for class-related objects

Because of all of this I decided to remake the entire thing from the
beggining in versionTwo.
Switching to class instead of bare code now allows for easier
modification of the code later, which is vital for when I would be
adding log system.

Switching to keyboard.add_hotkey now allows for the code to register the
hotkey even if the language is not english. Before this, the code would
crash whenever the language was anything but english

Aside from that, just many small optimizations to the code and changes
to the logic and names for easier reading